修正，待完善

+++ app.js
@@ -27,34 +27,75 @@ https.get(url, (res) => {
 			downloadImg('img/', movie.picUrl);
 		})
 		console.log(movies);
-		saveData('data/movies.json', movies);
+		saveData('data/', movies);
 	})
 }).on('error', (err) => {
 	console.log(err);
 })
 
 function saveData(path, movies){
-	fs.writeFile(path, JSON.stringify(movies), (err) => {
-		if (err) return console.log(err);
-		console.log('Data Saved');
+	fs.exists(path, (exists) => {
+		if (exists) {
+			// 文件夹存在
+			fs.writeFile(path + 'movies.json', JSON.stringify(movies), {'flag': 'w+'}, (err) => {
+				if (err) return console.log(err);
+				console.log('Data Saved');
+			})
+		}else {
+			// 文件夹不存在
+			fs.mkdir(path, (err) => {
+				if (err) return console.log(err);
+				fs.writeFile(path + 'movies.json', JSON.stringify(movies), {'flag': 'w+'}, (err) => {
+					if (err) return console.log(err);
+					console.log('Data Saved');
+				})
+			})
+		}
 	})
+	
 }
 
 function downloadImg(imgDir, url){
-	https.get(url, (res) => {
-		let img = '';
-		res.setEncoding('binary');
-		res.on('data', (data) => {
-			img += data;
-		})
-		res.on('end', () => {
-			fs.writeFile(imgDir + path.basename(url), img, 'binary', (err) => {
-				if(err) return console.log(err);
-				console.log('Image downloaded: ', path.basename(url));
+	fs.exists(imgDir, (exists) => {
+		console.log('exists: ', exists);
+		if (exists) {
+			// 文件夹存在
+			https.get(url, (res) => {
+				let img = '';
+				res.setEncoding('binary');
+				res.on('data', (data) => {
+					img += data;
+				})
+				res.on('end', () => {
+					fs.writeFile(imgDir + path.basename(url), img, 'binary', (err) => {
+						if(err) return console.log(err);
+						console.log('Image downloaded: ', path.basename(url));
+					})
+				})
+			}).on('err', (err) => {
+				console.log(err);
 			})
-		})
-	}).on('err', (err) => {
-		console.log(err);
+		}else {
+			// 文件夹不存在
+			fs.mkdir(imgDir, (err) => {
+				// if (err) return console.log(err);
+				https.get(url, (res) => {
+					let img = '';
+					res.setEncoding('binary');
+					res.on('data', (data) => {
+						img += data;
+					})
+					res.on('end', () => {
+						fs.writeFile(imgDir + path.basename(url), img, 'binary', (err) => {
+							if(err) return console.log(err);
+							console.log('Image downloaded: ', path.basename(url));
+						})
+					})
+				}).on('err', (err) => {
+					console.log(err);
+				})
+			})
+		}
 	})
+	
 }
-
